WIP: Expose a way to set "advanced" non-types scan options
This adds a new method to KuduScanner that allows to set
"advanced" scan options just by providing a pair of key
and value strings.

The main target for this feature is to allow Impala scanners
to specify that padding is required for UNIXTIME_MICROS columns
without "polluting" the api and in such a way that we can remove
support for this in the future.

This api could potentially be used to set "advanced" scanner hints
in the future.

WIP: because this needs a directed end-to-end test, even though
the wire-protocol itself has been tested in a previous patch.

Change-Id: I043b6514dc5fc307fc9c94eb41f3ae79796ba273